code cleanup and renaming
implement suggested changes (rename classes, move to method to util class)
