+++ bndtools.core/src/bndtools/editor/pages/BundleContentPage.java
@@ -27,7 +27,7 @@ import org.eclipse.ui.forms.widgets.Section;
 import aQute.bnd.build.model.BndEditModel;
 import aQute.bnd.build.model.clauses.ExportedPackage;
 import aQute.bnd.build.model.clauses.ImportPattern;
-import bndtools.editor.common.Buttons;
+import bndtools.editor.common.HelpButtons;
 import bndtools.editor.common.MDSashForm;
 import bndtools.editor.contents.BundleCalculatedImportsPart;
 import bndtools.editor.contents.GeneralInfoPart;
@@ -74,7 +74,7 @@ public class BundleContentPage extends FormPage {
 		// buttons top of form
 		IToolBarManager toolbar = scrolledForm.getForm()
 			.getToolBarManager();
-		toolbar.add(Buttons.HELP_BTN_BND_EDITOR);
+		toolbar.add(HelpButtons.HELP_BTN_BND_EDITOR);
 		toolbar.update(true);
 
 		Form form = scrolledForm.getForm();
