+++ bndtools.core/src/bndtools/editor/pages/ProjectRunPage.java
@@ -23,7 +23,7 @@ import org.eclipse.ui.forms.widgets.Section;
 import aQute.bnd.build.Workspace;
 import aQute.bnd.build.model.BndEditModel;
 import bndtools.central.Central;
-import bndtools.editor.common.Buttons;
+import bndtools.editor.common.HelpButtons;
 import bndtools.editor.common.MDSashForm;
 import bndtools.editor.project.AvailableBundlesPart;
 import bndtools.editor.project.RepositorySelectionPart;
@@ -144,7 +144,7 @@ public class ProjectRunPage extends FormPage {
 				.add(exportContrib);
 
 		form.getToolBarManager()
-			.add(Buttons.HELP_BTN_BND_EDITOR_RUN);
+			.add(HelpButtons.HELP_BTN_BND_EDITOR_RUN);
 
 		form.getToolBarManager()
 			.update(true);
