+++ bndtools.core/src/bndtools/utils/EditorUtils.java
@@ -2,8 +2,13 @@ package bndtools.utils;
 
 import java.lang.reflect.InvocationTargetException;
 
+import org.bndtools.core.ui.icons.Icons;
+import org.eclipse.jface.action.Action;
+import org.eclipse.jface.action.ActionContributionItem;
+import org.eclipse.jface.action.IAction;
 import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.jface.operation.IRunnableWithProgress;
+import org.eclipse.swt.program.Program;
 import org.eclipse.ui.IEditorPart;
 import org.eclipse.ui.IWorkbenchWindow;
 import org.eclipse.ui.forms.IFormPart;
@@ -35,5 +40,46 @@ public class EditorUtils {
 				return part;
 		}
 		return null;
+	}
+
+	/**
+	 * Creates a button with icon and tooltip.
+	 *
+	 * @param url
+	 * @param tooltipText
+	 * @return
+	 */
+	public static final Action createButton(String url, String tooltipText) {
+		Action btn = new Action("Help", IAction.AS_PUSH_BUTTON) {
+			@Override
+			public void run() {
+				Program.launch(url);
+			}
+		};
+		btn.setEnabled(true);
+		btn.setToolTipText(tooltipText);
+		btn.setImageDescriptor(Icons.desc("help"));
+
+		return btn;
+	}
+
+	/**
+	 * Creates a button with icon, text and tooltip.
+	 *
+	 * @param url
+	 * @param buttonText
+	 * @param tooltipText
+	 * @return
+	 */
+	public static final ActionContributionItem createButtonWithText(String url, String buttonText, String tooltipText) {
+		Action btn = createButton(url, tooltipText);
+		btn.setText(buttonText);
+
+		// the ActionContributionItem is required to display text below the icon
+		// of the button
+		ActionContributionItem helpContrib = new ActionContributionItem(btn);
+		helpContrib.setMode(ActionContributionItem.MODE_FORCE_TEXT);
+
+		return helpContrib;
 	}
 }
